+++ urls.py
@@ -18,11 +18,11 @@ urlpatterns = [
     url(r'^logout$',            logout, name='logout'),
     url(r'^$',                  TemplateView.as_view(template_name='index.html'), name='index'),
     url(r'^about$',             TemplateView.as_view(template_name='about.html'), name='about'),
-    url(r'^report/',            include('reports.urls')),
     url(r'^core/',              include('core.urls')),
     url(r'^bills/',             include('bills.urls')),
     url(r'^scan/',              include('scan.urls')),
     url(r'^contract/',          include('contract.urls')),
     url(r'^contrarch/',         include('contrarch.urls')),
+    url(r'^report/',            include('reports.urls')),
 ]
 urlpatterns += staticfiles_urlpatterns()